cambios

+++ AreaTriangulo/AreaTriangulo.cs
@@ -9,9 +9,9 @@ namespace AreaTriangulo
             double Base = 0, altura = 0, area = 0;
             Console.WriteLine("Calcula el area de un triangulo");
             Console.Write("Ingrese la base del triangulo: ");
-            Base = float.Parse(Console.ReadLine());
+            Base = double.Parse(Console.ReadLine());
             Console.Write("Ingrese la altura del triangulo: ");
-            altura = float.Parse(Console.ReadLine());
+            altura = double.Parse(Console.ReadLine());
             area = (Base * altura)/2;
             Console.WriteLine("El area del triangulo es : " + area);
         }
